requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
pymysql>=1.1,<2.0
